factors out wrapper into own file and rewrites initial data flow
